+++ example/command_runner/draw.dart
@@ -33,8 +33,8 @@ class SquareCommand extends Command<String> {
 
   @override
   FutureOr<String>? run() {
-    final size = int.parse(argResults?['size'] as String? ?? '20');
-    final char = (globalResults?['char'] as String?)?[0] ?? '#';
+    final size = int.parse(argResults?.option('size') ?? '20');
+    final char = globalResults?.option('char')?[0] ?? '#';
     return draw(size, size, char, (x, y) => true);
   }
 }
@@ -55,8 +55,8 @@ class CircleCommand extends Command<String> {
 
   @override
   FutureOr<String>? run() {
-    final size = 2 * int.parse(argResults?['radius'] as String? ?? '10');
-    final char = (globalResults?['char'] as String?)?[0] ?? '#';
+    final size = 2 * int.parse(argResults?.option('radius') ?? '10');
+    final char = globalResults?.option('char')?[0] ?? '#';
     return draw(size, size, char, (x, y) => x * x + y * y < 1);
   }
 }
@@ -93,8 +93,8 @@ class EquilateralTriangleCommand extends Command<String> {
 
   @override
   FutureOr<String>? run() {
-    final size = int.parse(argResults?['size'] as String? ?? '20');
-    final char = (globalResults?['char'] as String?)?[0] ?? '#';
+    final size = int.parse(argResults?.option('size') ?? '20');
+    final char = globalResults?.option('char')?[0] ?? '#';
     return drawTriangle(size, size * sqrt(3) ~/ 2, char);
   }
 }
@@ -116,9 +116,9 @@ class IsoscelesTriangleCommand extends Command<String> {
 
   @override
   FutureOr<String>? run() {
-    final width = int.parse(argResults?['width'] as String? ?? '50');
-    final height = int.parse(argResults?['height'] as String? ?? '10');
-    final char = (globalResults?['char'] as String?)?[0] ?? '#';
+    final width = int.parse(argResults?.option('width') ?? '50');
+    final height = int.parse(argResults?.option('height') ?? '10');
+    final char = globalResults?.option('char')?[0] ?? '#';
     return drawTriangle(width, height, char);
   }
 }