+++ test/parse_test.dart
@@ -83,6 +83,79 @@ void main() {
         var results = parser.parse(['--no-b']);
         expect(results['a'], isFalse);
       });
+
+      test('throws if requested as a multi-option', () {
+        var parser = ArgParser();
+        parser.addFlag('a', defaultsTo: true);
+        var results = parser.parse(['--a']);
+        throwsIllegalArg(() => results.multiOption('a'));
+      });
+    });
+
+    group('flag()', () {
+      test('returns true if present', () {
+        var parser = ArgParser();
+        parser.addFlag('verbose');
+
+        var args = parser.parse(['--verbose']);
+        expect(args.flag('verbose'), isTrue);
+      });
+
+      test('returns default if missing', () {
+        var parser = ArgParser();
+        parser.addFlag('a', defaultsTo: true);
+        parser.addFlag('b', defaultsTo: false);
+
+        var args = parser.parse([]);
+        expect(args.flag('a'), isTrue);
+        expect(args.flag('b'), isFalse);
+      });
+
+      test('are false if missing with no default', () {
+        var parser = ArgParser();
+        parser.addFlag('verbose');
+
+        var args = parser.parse([]);
+        expect(args.flag('verbose'), isFalse);
+      });
+
+      test('are case-sensitive', () {
+        var parser = ArgParser();
+        parser.addFlag('verbose');
+        parser.addFlag('Verbose');
+        var results = parser.parse(['--verbose']);
+        expect(results.flag('verbose'), isTrue);
+        expect(results.flag('Verbose'), isFalse);
+      });
+
+      test('match letters, numbers, hyphens and underscores', () {
+        var parser = ArgParser();
+        var allCharacters =
+            'abcdefghijklmnopqrstuvwxyz-ABCDEFGHIJKLMNOPQRSTUVWXYZ_0123456789';
+        parser.addFlag(allCharacters);
+        var results = parser.parse(['--$allCharacters']);
+        expect(results.flag(allCharacters), isTrue);
+      });
+
+      test('can match by alias', () {
+        var parser = ArgParser()..addFlag('a', aliases: ['b']);
+        var results = parser.parse(['--b']);
+        expect(results.flag('a'), isTrue);
+      });
+
+      test('can be negated by alias', () {
+        var parser = ArgParser()
+          ..addFlag('a', aliases: ['b'], defaultsTo: true, negatable: true);
+        var results = parser.parse(['--no-b']);
+        expect(results.flag('a'), isFalse);
+      });
+
+      test('throws if requested as a multi-option', () {
+        var parser = ArgParser();
+        parser.addFlag('a', defaultsTo: true);
+        var results = parser.parse(['--a']);
+        throwsIllegalArg(() => results.multiOption('a'));
+      });
     });
 
     group('flags negated with "no-"', () {
@@ -475,6 +548,13 @@ void main() {
         expect(args['define'], equals('2'));
       });
 
+      test('throw if requested as a multi-option', () {
+        var parser = ArgParser();
+        parser.addOption('a', defaultsTo: 'b');
+        var results = parser.parse(['--a=c']);
+        throwsIllegalArg(() => results.multiOption('a'));
+      });
+
       group('returns a List', () {
         test('with addMultiOption', () {
           var parser = ArgParser();
@@ -541,6 +621,111 @@ void main() {
           var results = parser.parse(['-h']);
           expect(results['help'], true);
           expect(() => results['test'], throwsA(isA<ArgumentError>()));
+        });
+      });
+    });
+
+    group('option()', () {
+      test('are parsed if present', () {
+        var parser = ArgParser();
+        parser.addOption('mode');
+        var args = parser.parse(['--mode=release']);
+        expect(args.option('mode'), equals('release'));
+      });
+
+      test('are null if not present', () {
+        var parser = ArgParser();
+        parser.addOption('mode');
+        var args = parser.parse([]);
+        expect(args.option('mode'), isNull);
+      });
+
+      test('default if missing', () {
+        var parser = ArgParser();
+        parser.addOption('mode', defaultsTo: 'debug');
+        var args = parser.parse([]);
+        expect(args.option('mode'), equals('debug'));
+      });
+
+      test('allow the value to be separated by whitespace', () {
+        var parser = ArgParser();
+        parser.addOption('mode');
+        var args = parser.parse(['--mode', 'release']);
+        expect(args.option('mode'), equals('release'));
+      });
+
+      test('do not throw if the value is in the allowed set', () {
+        var parser = ArgParser();
+        parser.addOption('mode', allowed: ['debug', 'release']);
+        var args = parser.parse(['--mode=debug']);
+        expect(args.option('mode'), equals('debug'));
+      });
+
+      test('do not throw if there is no allowed set with allowedHelp', () {
+        var parser = ArgParser();
+        parser.addOption('mode', allowedHelp: {
+          'debug': 'During development.',
+          'release': 'For customers.'
+        });
+        var args = parser.parse(['--mode=profile']);
+        expect(args.option('mode'), equals('profile'));
+      });
+
+      test('returns last provided value', () {
+        var parser = ArgParser();
+        parser.addOption('define');
+        var args = parser.parse(['--define=1', '--define=2']);
+        expect(args.option('define'), equals('2'));
+      });
+
+      test('throw if requested as a multi-option', () {
+        var parser = ArgParser();
+        parser.addOption('a', defaultsTo: 'b');
+        var results = parser.parse(['--a=c']);
+        throwsIllegalArg(() => results.multiOption('a'));
+      });
+
+      group('returns a List', () {
+        test('with addMultiOption', () {
+          var parser = ArgParser();
+          parser.addMultiOption('define');
+          var args = parser.parse(['--define=1']);
+          expect(args.multiOption('define'), equals(['1']));
+          args = parser.parse(['--define=1', '--define=2']);
+          expect(args.multiOption('define'), equals(['1', '2']));
+        });
+      });
+
+      group('returns the default value if not explicitly set', () {
+        test('with addMultiOption', () {
+          var parser = ArgParser();
+          parser.addMultiOption('define', defaultsTo: ['0']);
+          var args = parser.parse(['']);
+          expect(args.multiOption('define'), equals(['0']));
+        });
+      });
+
+      test('are case-sensitive', () {
+        var parser = ArgParser();
+        parser.addOption('verbose', defaultsTo: 'no');
+        parser.addOption('Verbose', defaultsTo: 'no');
+        var results = parser.parse(['--verbose', 'chatty']);
+        expect(results.option('verbose'), equals('chatty'));
+        expect(results.option('Verbose'), equals('no'));
+      });
+
+      test('can be set by alias', () {
+        var parser = ArgParser()..addOption('a', aliases: ['b']);
+        var results = parser.parse(['--b=1']);
+        expect(results.option('a'), '1');
+      });
+
+      group('mandatory', () {
+        test('parse successfully', () {
+          var parser = ArgParser();
+          parser.addOption('test', mandatory: true);
+          var results = parser.parse(['--test', 'test']);
+          expect(results.option('test'), equals('test'));
         });
       });
     });